+++ CHANGELOG.rst
@@ -75,6 +75,10 @@ In development
   kombu, six, appscheduler, passlib, python-gnupg, semver, paramiko, python-keyczar, virtualenv).
 * Improve performance of ``GET /executions/views/filters`` by creating additional indexes on
   executions collection
+* Add support for default values and dynamic config values for nested config objects. (new feature,
+  improvement)
+* Throw a more user-friendly exception if rendering a dynamic configuration value inside the config
+  fails. (improvement)
 
 2.0.1 - September 30, 2016
 --------------------------